pyproject.toml
[tool.poetry]
name = "huawei-sun2000l-mqtt"
version = "0.1.0"
description = ""
authors = ["Jordi Bofill <jordi.bofill@gmail.com>"]
readme = "README.md"
packages = [{include = "huawei_sun2000l_mqtt"}]

[tool.poetry.dependencies]
python = "^3.8"
pymodbus = "^2.5.3"
asyncio = "^3.4.3"
continuous-threading = "^2.0.5"
langcodes = "^3.3.0"
# paho-mqtt = ">=1.6.1,<1.7.0"
paho-mqtt = "^1.6.1"
tabulate = "^0.8.10"


[tool.poetry.group.dev.dependencies]
ipython = "^8.5.0"
pytest = "^7.1.3"
Babel = "^2.10.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
